Set window background color (#27)
Makes window resizing feel slightly less janky.

I don't know why `exportOnlyLocals` is necessary here but not in webpack.renderer.config.js. In the renderer we are able to `import colors from 'colors.module.scss';` and it works fine, but when I tried that in the main process it imported as `{ locals: { ...colors... } }` instead of just `{ ...colors... }`.