Speed up BZ2 decompression with WebAssembly (#346)
Adds a small library in the same vein as https://github.com/cruise-automation/wasm-lz4, for bzip2 decompression.

Time to fully preload data from a 62MB bag compressed with bz2:
- v0.2: ~14sec
- this branch with `yarn build:prod && yarn package:darwin`: ~5sec
(Tested by watching the clock, dropping a bag file and seeing how long it takes for the playback bar to fill up.)

Sadly this is still much slower than the same bag with lz4 compression, which takes ~1 sec or less to parse.