Fix Webpack build warnings from bundling TypeScript (#546)
TypeScript has some conditional code that runs when it is executing in a nodejs environment, where it can do `require("perf_hooks")` and `require(some dynamic string)`. Webpack is unhappy with these, but they would never run in the browser anyway, so we can just disable them by string-replacing the requires out.