Reduce Panel re-renders and add perf indicator (#621)
There were a couple of overly broad useSelectors that caused each Panel to re-render whenever the layout or saved props of any other panel changed.

Fixed these and added a React.Profiler around each panel to show a little perf indicator in the corner in dev mode, which shows the actualDuration of the last render as well as the number of renders.