Fix provider remount when toggling fullscreen (#819)
A fullscreen toggle results in a re-render of the App. This caused
all of the providers to unmount and remount which re-initializes a lot
of app state and variables we do not want to re-initialize to just resize
the window.

The remount happened because of how AllProviders was declared. A new functional
component of AllProviders was created within the App render function. This results
in the old component being unmounted and fully replaced on every render.

Additionally a similar problem was happening within ThemeProvider. A new component
declaration was created on every render and passed to the _as_ property of FluentThemeProvider. This resulted in all children under FluentThemeProvider being unmounted
and recreated.