Sidebar (#816)
Adds a new Sidebar to the app, and moves the Add Panel flow, Variables menu, and Settings into the sidebar.

The sidebar uses another react-mosaic-component to allow resizing and separate it from the rest of the app content, but it is not part of the actual saved layout.

Future work will be to move individual panel settings into the sidebar (initially part of this PR but temporarily reverted in 9ab66b9) as well as the layout selector, and finally clean up the window title bar height and appearance.