Merge redundant loadLayout and importPanelLayout actions (#1001)
These both had exactly the same effect. We don't need both actions to exist.

The skipSettingLocalStorage: true feature was not used for this action, so it is removed from the payload. It is still used in some other actions, but may be 🔪able — it's not a configurable option, just a statement that certain actions should not affect localStorage. And these actions already store their state outside the persistedState, so in fact they can't affect localStorage — all the code currently does without the skip flag is some redundant serialization work to set the same value in localStorage.