Fix panel settings being incorrectly overwritten (#1089)
useCurrentLayoutSelector had a bug where a change in the selector only (but no change in the panel config) would not trigger a re-render and update the result. This was triggered by panel settings sidebar, which changes the selector when a panel is selected or deselected. This led to a stale or incorrect config (such as from another panel) being merged with the default config, and the wrong settings being saved (thanks @esthersweon for catching this!).

Fixing the selector bug by adding another setState would be simple, but I wanted to avoid unnecessary renders and also fix a race condition between subscribing and state updates happening in an effect, so I went for a more involved change. (A similar race condition may occur in a couple other places in the app, `useContextSelector` and `useSelectedPanels`, but those can be tackled later in separate PRs.)