Add Quick Look preview for bag files (#1187)
Integrate with the macOS Quick Look extension API to provide a preview summary for bag files. The preview shows up in Finder and the Dock, and can be invoked by other applications.

This PR contains a new webpack entrypoint in the desktop build for the preview page, and an electron-builder afterPack script to bundle the app extension (beyond just copying files, this requires a codesign step to properly sandbox the extension, which means the build step will only work on macOS). The native code implementing the app extension is at https://github.com/jtbandes/quicklookjs.