Don't render chart data if there are no bounds for downsampling (#1188)
When the plot is initially mounted the bounds are not known. This
resulted in trying to render the entire dataset rather than a downsampled
version. For large datasets, this caused a blank chart while the
dataset was transferred to the worker and rendered which can take seconds
for large datasets. This occurs when adding panels to alter the layout.

This change updates the downsample behavior to return no datapoints
if the chart bounds are not known (during initialization) so that regular
downsampling can happen and avoid render delays.