Revert "Reset all panels when player changes. (#1008)" (#1228)
This reverts commit de0beac2c8ac11aec180e671dfa8be528c5006aa.

Reverts #1008.

This PR broke the app so that bags cannot be loaded by drag/drop or double click, without triggering either an ReactDndHtml5Backend error or an infinite re-render loop.

Theoretically all components should be designed so that they are able to unmount and remount without triggering this kind of error. It merits further investigation, but for now we need the app to work.

No user impact (the original PR also had no intended user impact)