Fix some no-explicit-any in 3D panel (#1257)
**User-Facing Changes**
none

**Description**
Many of the type definitions around 3D panel, World, and Interactions are incomplete/incorrect which requires some sketchy casts. I don't think this is a small undertaking to tackle. So for now I left several `as` casts in.

For some cases, added an eslint-disable and filed #1255 to track some additional refactoring work