Enable and fix `filenames/match-exported` (#1417)
Unfortunately, this rule is no longer maintained. However, it is functional (although reports error locations in a way that requires a suppression comment to be at a strange location) — and there doesn't seem to be much community consensus yet on a replacement (see benmosher/eslint-plugin-import#662).

Related rule for imports: benmosher/eslint-plugin-import#325