Change: signin flow uses a dialog (#1548)
A dialog feels more natural for the device code flow by helping
the user focus on the signin process rather than hiding it in the sidebar.

This change also adjusts the org selection behavior. If the user does not have
any orgs, they are presented with a notice to have their org admin invite them
or to create an org on the website.

Fixes: #1491