Fix: download image at original size (#1591)
The image panel provides a context menu option to download the shown
image. This option was exporting the entire canvas. Since we upsize
the canvas for the viewport (to provide for smooth marker rendering),
the resulting image file was not the original and contained transparent area.

This change fixes the image download by re-rendering the image onto a new canvas
that is appropriately sized for the original image. This new canvas is then
used to create the image file.