Change: remove in-app titlebar (#1677)
On Windows and Linux we are not able to overlay native window controls
into the app and omit the system titlebar. This creates a "dual bar"
appearance with the system titlebar and our in-app toolbar. This dual
bar takes up screen space with little value - only the current connection string and problems are placed there.

This change removes the in-app toolbar and leaves the native window
titlebar and controls. It also moved the problems into the connection
sidebar pane and adds badges to the connection sidebar icon when there are problems with the connection.