Don't load a new Welcome layout when deleting the last layout (#1721)
**User-Facing Changes**
None worth mentioning in release notes

**Description**
Fixes #1545, closes #1570, closes #1575

There used to be 3 places where a welcome layout might be created: (a) on first app launch, (b) on every app launch when the list of layouts is empty, and (c) when deleting the last layout. (b) was removed in #1679. This PR removes (c), so there is now only 1 place that loads the welcome layout.

Deleting the last layout now results in _no_ layout being selected, rather than auto-loading the welcome layout.