Fix: save any unsaved layouts when writing to layout manager (#1753)
This change updates CurrentLayoutProvider to save any unsaved layouts
when it writes to the layout manager. Previously, only the latest changed
layout would be saved. This was fine in practice but if a user did manage
to make a change to a layout, switch to another layout, and make a change
in under 1 second (the save interval), their older changes would not be saved.