Fix: remove sendNotification from PlaybackControls slider (#1805)
Instead of notifying when the max/min range is inverted, the slider will
render nothing and log a console error.

This change updates the slider component from a class component to a functional
component. It also removes the document events module and listens to events on
the slider div directly.

Fixes #1804