Fix: always keep dataset endpoints when downsampling (#1830)
**User-Facing Changes**
Improved plot downsampling algorithm to avoid introducing gaps with adjacent datasets.

**Description**
The downsampling algorithm would sometimes drop the last datum in a dataset, which tends to introduce visual artifacts in the State Transtions panel because most of its datasets are immediately adjacent to each other.

Change the downsampling algorithm to always keep the last point (and add unit tests).