Disable "Update team layout" from unsaved changes prompt while offline (#1879)
**User-Facing Changes**
None (feature flagged)

**Description**
Disable the radio button to save a team layout from the unsaved changes prompt while offline.

To accomplish this without adding a useLayoutManager dependency to the UnsavedChangesPrompt component itself (which would require a fake layout manager in storybook), I reworked the `useUnsavedChangesPrompt` hook so instead of using the ModalHost imperative API (which would make it hard to dynamically update the disabled state), the hook manages the state itself and **returns** the react element for the prompt which can be rendered anywhere in the tree by the caller.