Fix: prevent rpc calls before chart initialized (#1888)
In some instances the hover handler would attempt to dispatch rpc
calls prior to the chart initialization. This would result in errors
in the worker since a chart with the specific id does not yet exist.

This change fixes this issue by delaying the availability of the rpcSend wrapper until the chart is initialzed in the worker.

Fixes: #1873