Don't show sign in prompt and account pane if ConsoleApi is not present (#1905)
**User-Facing Changes**
None (feature flagged)

**Description**
The Account tab of the sidebar and the sign-in prompt in the LayoutBrowser were previously shown when the Team Layouts feature flag was enabled. Now they are additionally gated on the presence of ConsoleApi, so when the feature flag is removed, these features will remain hidden if no ConsoleApi is provided.

Resolves https://github.com/foxglove/studio/issues/1901.