Add mcap support to Quick Look extension (#1981)
**User-Facing Changes**
None (mcap is still experimental)

**Description**
For ease of development, add support for mcap to the Quick Look extension. The preview is slow for large files because it parses the whole file rather than using random access (since we haven't implemented indexes yet). Still it's better than nothing.

Includes a couple changes to the mcap parser: ability to disable chunk CRCs, and polyfill for DataView.getBigUint64.