Change "log data to console" to "copy" for RawMessage arrays (#2016)
Rather than showing a "Log data to console" button for arrays in
the RawMessage panel, show a copy icon. When the user clicks the icon,
the array data is copied to the clipboard.

This makes more sense for a desktop app where the notion of a "browser console" is not present. Even in the web it isn't that user friendly.