Improve UX when no layout is selected (#2034)
**User-Facing Changes**
Improved the user experience for adding panels in rare cases when no layout is currently selected.

**Description**
Fixes https://github.com/foxglove/studio/issues/2007 by adding a link to open the Layouts sidebar when no layout is selected, and also hiding the content of the Add Panel and Panel Settings sidebars to avoid confusion.