Attempt to cancel in-progress layout sync when remounting (#2048)
**User-Facing Changes**
None

**Description**
Thread AbortSignal through the layout sync logic to improve ability to interrupt an ongoing sync.

The root cause of certain sync errors is that the tree is remounting when the app initially loads, but this seemed like something we could do to be more defensive against problems.

See also: #2051