Don't re-initialize layout storage if current user is rebuilt with same user id (#2051)
**User-Facing Changes**
Fixed an issue where some spurious errors would be displayed when syncing layouts after the app first opened.

**Description**
A new ConsoleApiRemoteLayoutStorage was being initialized whenever the currentUser object changed, even if the user id was the same, although it only depended on the user id.

The ConsoleApiCurrentUserProvider unfortunately triggers this behavior on every app launch because it always calls setCachedCurrentUser even if it was just validating the cached user:
https://github.com/foxglove/studio/blob/44212a8aaab9896b5ae30bf212c5eb3234034890/desktop/renderer/components/ConsoleApiCurrentUserProvider.tsx#L68

Change the useMemo dependency to be the id itself and add a unit test.

See also: #2048