Remove feature branch support from 3d panel (#2095)
This was a webviz/cruise specific feature which was not documented nor supported
for anything but the bag workflow with the 3d panel and raw message panel.

We recently removed support for loading multiple ros1 bag files as the feature was not documented and not well supported across the app. This is cleanup from that for the various bits of code in 3d panel.

The config shape does change so a forward migration is introduced.