Ignore current seek when a new seek occurs in RandomAccessPlayer (#2231)
In RandomAccessPlayer when a seek is already loading messages, new seek requests are queued (queue size of 1). Prior to this change, when a seek request arrived while an existing seek request was in-flight, the existing request would finish, emit state, and then the new request would start.

Prior to emitting state, the in-flight seek request would set the nextReadStartTime which would override the nextReadStartTime of the latest seek request. Even if the seek requests were for the same time, the first request would alter the next read time of the subsequent request.

This change updates the RandomAccessPlayer logic to detect if a seek request occured while another was in-flight and avoid emitting a state update for the now-stale seek request.

Fixes: #1712