Workaround for Change Panel menu positioning (#2234)
**User-Facing Changes**
Fixed a visual bug with the alignment of the "Change panel" menu.

**Description**
Fixes #2205 by making the menu full height even when it is filtered.

The underlying problem is that fluentui's Callout doesn't reposition itself when the content size changes (https://github.com/microsoft/fluentui/issues/18839). This appears to have been addressed in https://github.com/microsoft/fluentui/pull/18887 but later reverted and the original issue is still unresolved.