WebSocket player: clear message queue when time goes backward (#2235)
**User-Facing Changes**
None (feature flagged)

**Description**
A "Bag went back in time" error would be displayed if time looped when the app was in the background. This is because the processing of messages is blocked on a requestAnimationFrame, and the logic that incremented lastSeekTime was only looking at the last message in the queue on each emit. This PR changes the logic to check each message as it arrives, and if time goes backwards, we empty out the queue immediately.

This will help with unbounded memory growth for a server that uses looping playback, but we can still have unbounded growth for a monotonic server when the app is in the background. We should probably implement some kind of max queue size and start throwing away old messages if they are not getting processed.