Remove box-shadow from SidebarContent (#2315)
The box-shadow was a hack around a hack for mosaic tiles overlapping with mosaic splitters. The mosaic tile has was removed in a separate PR. This resulted in the remaining box-shadow hack hiding the border for the sidebar icon stack.

This change removes the box-shadow hack for SidebarContent so the sidebar icon stack border is shown again.

Fixes: #2313