Update player with global variables on initialization (#2453)
During player initialization the global variables would be cached to the current value from the layout. If a layout was still loading these variables would be empty. When creating a player, the constructor would receive these empty cached values rather than the latest values.

This change updated the PlayerManager logic to use the latest global variable value when constructing a player. Care is taken to avoid constructing players when the variables change.

Fixes: #2251