Avoid creating a new ConnectionSidebar component (#2465)
The ConnectionSidebar component is dynamic to support the enableOpenDialog feature flag. This component was created in a useMemo that had player problems as a dependency. As a result, whenever a player problem occured the entire component would be re-created and re-mounted.

This change moves the dynamic sidebar component into its on useMemo so it is created once rather than on every player problem change.

This change also moves some component logic into a new DataSourceSidebar folder to tidy up Workspace.

Fixes: #2434