Remove support for LinedConvexHulls (#2490)
* Remove support for LinedConvexHulls

Remove support for toggling "Line marker click events override", which had the description: "Optionally allow treating line markers as polygons, so that clicking inside the lines in the line marker selects the marker. The default behavior for line markers requires the user to click exactly on the line to select the line marker. This option can reduce performance and will not work on instanced line markers (those with type: 108)"

* Remove quickhull3d runtime dependency