Update rosbridge player to re-subscribe to all topics (#2758)
Previously, new panel subscriptions to existing latching topics would
never receive the latched message. This happened because the RosbridgePlayer would ignore subscriptions to topics it was already subscribed to but would not send any previous messages on these topics to the new subscribers.

This change updates RosbridgePlayer logic to perform an unsubscribe/resubscribe cycle whenever the topic subscriptions change. This allows the underlying rosbridge server to process the subscription again and send any latched messages.

See: #2343