Update SceneBuilder to check for required PointCloud fields. (#2766)
PointCloud rendering requires field descriptions for x, y, and z fields. This change updates SceneBuilder to check for these fields prior to adding the point cloud to the collector. This allows the UI to display a topic error for the malformed topic better informing the user to errors with their data.

Fixes: #2759