Avoid automatically highlighting the first entry in the Add Panel List (#2771)
When showing the add panel list, no entries are highlighted by default. If a user types in the search box, the first entry is highlighted - allowing for quickly filtering and pressing "enter" to add the panel.

When the user clears the search box, highlighting is removed.

Keyboard navigation works as before - with "key down" moving down the list and "key up" moving up. Both wrapping around at each end of the list.

Fixes: #1589