Inject the last message on a topic on new subscriptions (#2778)
**User-Facing Changes**
When adding a panel which needs to subscribe to an infrequently publishing topic (i.e. /tf_static), the new panel will receive the message.

Previously only the first panel to subscribe would receive the message.

**Description**
This change updates the MessagePipelineProvider to add logic for injecting the last message on a topic to new subscribers of the topic. The provider tracks when new panels subscribe, via setSubscriptions calls, and makes a note on any new topics requested by the panel. When a new player state arrives, the pipeline checks any of these new topics and sends the last message on any of the topics to the newly subscribed panels.

To avoid sending the last message to previously subscribed panels, the provider adds a new context field `messageEventsBySubscriberId`. `useMessageReducer` and `PanelExtensionAdapter` read their message events from these per-subscriber queues rather than from a single `messages` firehose. This mechanism allows the provider to target message event updates to specific subscribers.

Once a new subscriber has received a backfilled message, it is cleared from the new subscriber list to avoid it receiving any further backfill messages. Only the last message on a topic is provided as backfill.

Fixes: #2343

Co-authored-by: Jacob Bandes-Storch <jacob@foxglove.dev>