Show context menu when clicking overlapping markers (#2845)
**User-Facing Changes**
Clicking overlapping markers in the 3d panel shows a menu to select the specific marker.

**Description**
When clicking on overlapping markers in the 3d panel, display a menu to select the specific marker to inspect. Prior to this fix, clicking overlapping markers resulted in one being selected with no way to select others.

This functionality was broken by https://github.com/foxglove/studio/pull/2617

Fixes: #2844