Add DialogHostIdContext for controlling modal attachment (#2958)
Allows users to specify the id of a DOM element to be used as the parent element for modals.

**User-Facing Changes**
None.

**Description**
Enables the inclusion of a provider to control which DOM node to which modals will be appended. If not provided, the default behavior of being added to the end of the document is retained.

Sample: <DialogHostIdContext.Provider value={"my-cool-div"} />