Sort headerStamp series in the plot panel (#2976)
The contract studio has when providing messages in "receive time" order is that the messages are provided in increasing receive time order. However, within a single receive time, there are no further guarantees about message ordering. If there are multiple messages at a receive time with different header stamps, these could cause the plot to render lines "out-of-order" when an earlier header stamp appeared in the dataset after a later header stamp (both at the same receive time).

This change fixes the "out-of-order" header stamp issue by sorting the dataset when using the header stamp method for a series. This ensures that the data is visualized in increasing header stamp order.