Bag playback with forward and reverse iterators (#2915)
**User-Facing Changes**
Add a new feature flag for "Experimental bag player".

When enabled, using a bag data source (remote or local) will use a new player with new seeking capabilities. Seeking to any location will load the last message on each subscribed topic into panels. This lets panels display the last message on a topic even if it occurred much earlier in the file.

**Description**
This change introduces a new IterablePlayer which uses iterators for message reading and preloading. When seeking, the player uses a reverse iterator to backfill the latest message on all subscribed topics and emit this to the message pipeline. When playing forward the player uses a forward iterator to grab latest messages for the _tick_ and also for the preloaded blocks. Similar to the RandomAccessPlayer, the player accepts pluggable instances which implement message iteration (IIterableSource) for an underlying data source. This change adds a BagIterableSource for handling local and remote bag files. 

The existing RandomAccessPlayer and `getMessages` interfaces were unable to support arbitrary reverse iteration and were designed around fetching messages for pre-defined start/end ranges.

Fixes: #707