Improve error messages for console/DP API auth errors (#3093)
**User-Facing Changes**
Improved error messages when attempting to access data from Data Platform when not logged in or logged in to the wrong organization.

**Description**
Fixes #3079

Added a custom human-readable message to 401/403 errors.

Pipe through the `{"error":"..."}` message from the error response in other cases (we were already passing through a `message` field, but not an `error` field).