Fix jumping dropdown menu in Diagnostic panels (#3153)
When using the topic selection menu in the Diagnostic panels (detail and summary), the context menu would jump around on the screen during playback or new messages.

The panel toolbar disappears when the user mouses away from the panel. Since a context menu is considered "outside" the panel the disappearing toolbar removes the icon from acting as the anchor point.

This change fixes the jumping behavior by using a static anchor for the menu that is outside of the disappearing panel toolbar.

Fixes: #3097