Recursively search ROS_PACKAGE_PATH in findRosPackage (#3157)
ROS package:// urls follow the convention of `package://<package name>/<path to file>`. In ROS projects, it is common to have a monorepo with several packages folders. This folder would be in the ROS workspace. When setting the ROS_PACKAGE_PATH environment variable or app setting to point to the workspace, the correct behavior for finding packages is to look recursively at the ROS_PACKAGE_PATH location and any locations below it to find the package.

This change updates our package lookup logic to use a recursive search on the ROS_PACKAGE_PATH. This allows finding packages which are within parent folders under the ROS_PACKAGE_PATH.

Fixes: #3088