MemoryCacheDataProvider: fix crash by increasing readAheadBlocks to minimum of 2 (#3205)
**User-Facing Changes**
Fixed a bug where seeking in large files would occasionally crash.

**Description**
Fixes https://github.com/foxglove/studio/issues/3099

When the block sizes are large enough (for large files), `_readAheadBlocks` may be 1. However, depending on the data range requested by the player during `_tick()`, it is always possible for the request range to span 2 blocks if it falls near a block boundary.

We considered simply removing the failing assertion. However, if we removed the assertion rather than increasing the `cacheSize`, it would result in `getNewConnection` returning a smaller range than initially requested, which would lead to MemoryCacheDataProvider violating its API contract (that `getMessages` should always return _all_ messages that were requested).

Some confusion resulted from the naming of the `cacheSize` variable, which is not exactly a hard limit on the cache size, but a hint as to how much data to load with each request, so this PR renames the variable to `maxRequestSize`.