Allow 100 milliseconds for IterablePlayer to backfill (#3298)
When seeking, the IterablePlayer would immediately emit a message with the new seek time. This would show an updated seek location in the scrub bar but would also clear all messages from the screen since there were no messages loaded for the seek time yet.

This change gives the backfill 100 milliseconds to emit state before emitting the new seek time. For local data this allows users to scrub continuously and see the scene or messages rather than a black screen.

Fixes: #3297