Change Diagnostics panel names to "ROS Diagnostics" (#3388)
**User-Facing Changes**
Changed the names of "Diagnostics – Summary" and "Diagnostics – Detail" panels to reflect that they are ROS-specific.

**Description**
We don't intend to create Foxglove equivalent messages under @foxglove/schemas for these panels in their current form, because they are so ROS-specific. In the future, if there is user demand, we may consider a redesign of these panels so that they are more widely applicable.

Relates to https://github.com/foxglove/roadmap/issues/10