Add `simpleGetMessagePathDataItems` which doesn't require topics & datatypes (#3609)
**User-Facing Changes**
None

**Description**
Adds a `simpleGetMessagePathDataItems()` which performs the same path-execution functionality as `getMessagePathDataItems()` but without requiring topics and datatypes as input. This will be usable from extension panels (although we should remove the "RosPaths" stuff before we actually publish it with `@foxglove/studio`).

This is required for #2003 / #1985 in order to use the `input: 'messagepath'` from extension panels.