Fix foxglove.FrameTransform handling in 3d panel (#3611)
The 3d panel subscribes to transform topics based on datatype. The `foxglove.FrameTransform` datatype was not listed in the TF_DATATYPES constant which meant the 3d panel never subscribed to topics of that datatype.

This change fixes this behavior by adding the foxglove FrameTransform datatypes to the TF_DATATYPES constant.

Additionally, useTransforms is updated to take the correct code paths for foxglove FrameTransforms vs ROS transforms and mark the transforms as updated.