Distinguish between un-indexed and empty mcap files (#3768)
**User-Facing Changes**
Users of the experimental mcap player are given a more detailed error message if their mcap file cannot be opened with the indexed reader.

**Description**
Mcap files may have summary records but be missing chunkindex records (as a result of not having any messages). We distinguish between files with no summary information and those that might have some summary information but are missing the summary information we need.

Fixes: #3731