NodePlayground: Use empty datatype definition for unknown datatypes (#3818)
When a topic references a datatype that is not yet known to the data source, the types library would fail to compile and run. The user would need to manually add a subscriber (some panel) for a topic of the missing datatype to unblock node playground.

This changes fixes this requirement by using empty datatype placeholders for unknown datatypes so the user script compiles and runs.

Fixes: #3784