Never show message frequency when there is only a single message (#3866)
**User-Facing Changes**
The data source sidebar no longer shows a frequency for topics that contain only a single message.

**Description**
If we have only one message or zero duration, don't presume to calculate a frequency. Notice that the `numMessages < 2` condition was already there, but it was only used if the data source was able to provide firstMessageTime/lastMessageTime info.

Fixes https://github.com/foxglove/studio/issues/3732