Limit ROS Diagnostic panels to DiagnosticStatusArray topics (#3874)
Prior to this change, the Diagnostic panels would always subscribe to the /diagnostics topic regardless of the datatype. This resulted in the panels crashing when a user sent a message that is not a DiagnosticStatusArray on this topic.

This change fixes the panel logic to avoid subscribing to the topic unless it publishes a supported datatype.

Fixes: #3785