Add buffering and caching for IterablePlayer playback (#3772)
**User-Facing Changes**
When playing using the experimental players, the playback will cache and buffer leading to a smoother and faster playback experience.

**Description**
Buffering and caching are implemented via two wrapper classes. Each class implements IIterableSource and exposes a messageIterator and getBackfillMessages.

The BufferingIterableSource does read-ahead buffering of an IIterableSource. This read-ahead happens in parallel to the message reading. Messages are produced into a read-ahead buffer by the producer "process" and consumed in the messageIterator.

The CachingIterableSource does memory caching on messages produces by an IIterableSource. It's messageIterator tries first to read from its memory cache, and when there are no more cached messages, falls-back to the underlying IIterableSource. When it falls-back to reading from the source, these results are cached in memory for future reading. These results are also available to getBackfillMessages.

CachingIterableSource organizes the iterator result cache into "blocks". A block has an array of cached iterator result items within a start and end time. The start and end time are inclusive of the time span the block covers. When a block grows too large in size (based on tracking MessageEvent#sizeInBytes), a new block is started. When the entire cache is grows too large, the oldest used blocks are purged.

Fixes: #3749