Delete unlimited memory cache setting (#3904)
**User-Facing Changes**
The unlimited memory cache setting is removed.

**Description**
This setting is not practically useful since the app is likely to crash or slow-down substantially when trying to use too much memory in a browser environment. This is an old setting carried over from webviz and not something we want to encourage for Studio users right now.

Fixes: #3903