Improve full preload handling for iterable player (#3934)
**User-Facing Changes**
Better handling of preloading topics for users on experimental players

**Description**
The BlockLoader is responsible for loading subscriptions with the "full" subscriber option. These are topics which a panel has indicated it wants all the messages across the entire data source range. When there is not enough memory to hold the entire dataset, the loader attempts to prioritize showing the data at and after the current time.

The existing behavior of BlockLoader had some drawbacks:

- When switching from idle -> play (and vice versa) the loader would stop and restart.
- When loading blocks the loader would not consider the position of the playhead for which blocks to evict vs keep which resulted in evicting blocks relevant to the current playhead.

This change addresses these issues by making the loader a "background" task of the iterable player. Rather than starting/stopping the loader during state changes, the iterable player updates the loader with the current time and the current full subscriber topic set. The loader uses these inputs to manage the full preload. When these inputs allow, the loader keeps loading rather than starting and stopping. When these inputs change in a way that the loader needs to start loading another region (i.e. seeking to an un-loaded region), the loader stops and restarts itself to always ensure it is loading the region with the playhead.

This change adds additional tests to verify that the loader loads the expected regions under playback and seeking.