Add RosDb3WorkerIerableSource (#3996)
**User-Facing Changes**
None

**Description**
The RosDb3WorkerIterableSource implements an IterableSource interface for reading ros2 .db3 files.

Reading ros2 .db3 files must be done in a worker to leverage the FileReaderSync interface for sqlite. We need this interface to read files larger than can fit in memory. To support using a worker, this change uses the comlink library and a custom transfer handler for serializing/deserializing the async message iterator.

Fixes: #3983