Add a memory use indicator option to experimental settings (#3748)
**User-Facing Changes**
Users can enable a memory use indicator in the app settings. The indicator shows the used javascript memory as a percent and used vs free as bytes in a tooltip.

**Description**
The memory use indicator shows used js heap memory relative to total js heap memory. This is useful in understand the memory impact of various caching strategies.