Add optional playUntil method to Player interface (#4045)
**User-Facing Changes**
Forward seek and backwards seek step forward and backwards the same amount of time.

**Description**
Our current behavior for forward seeking (via arrow key or seek button), is to "play" the data forward at least the seek amount. The purpose of this behavior is to ensure replay of stateful messages (like tf and markers) so pressing "seek" from the start of the file results in the same playback as playing from the start of the file.

Unfortunately, the way we implemented this feature was to start playback and attempt to pause playback after the seek target was reached. This behavior would result in a forward seek moving forward different and variable amounts of time since the "pause" would be a best effort invocation after some frames of messages.

This change removes this workaround play/quick-pause behavior in favor of first-class support within the player. The player can optionally implement a playUntil method. If available, seeking forward will invoke this method instead of regular seek. The player can implement playUntil however it wants. For IterablePlayer, the implementation emits all the messages in one "tick" to make forward seeking a small amount appear like a jump.

Fixes: #3980
Fixes: #3872
Fixes: #4059