Fix missing messages from CachingIterableSource (#4126)
**User-Facing Changes**
Fixes messages drops when playing data from data sources which contain multiple messages at the exact same log time (or receive time).

**Description**
When reading from the cache, the CachingIterableSource uses a binary search to deterline where to start reading in a cache block. If a block contained multiple messages at the exact same time as the search parameter, the search algorithm would return a valid index for any time matching the search parameter not necessarily the first occurrence in the block. By not returning an index for the first occurrence, some messages would be skipped during playback.

This change updates FindStartCacheItemIndex to handle this case. If an exact match is found in the cache block for the target time, the new logic iterates the block backwards until we find the first occurrence of the target time.

Fixes: #4091