Fix variable sidebar (#4272)
The core of these issues is the `key` field within each <Variable> element in the Variable sidebar react component. This key is the list index which meant that when a variable in the middle of the list was removed and the render was re-creating the components from the `map` over the variables, the index made react think the particular item is the same key.

It would pass in the new name and props but because the editableName state was only set to the `name` prop once, the element displayed the wrong (stale) name.

This change updates the keys to key off the variable names and only updates the variable name once the user has blurred from the input. This allows the user to freely iterate on the variable name even through duplicate names and only updates the name at the end if it isn't in conflict.

Fixes: #4216