Call pauseFrame/resumeFrame in PanelExtensionAdapter (#4233)
When handling context updates it is important to call pauseFrame to indicate that the panel will perform some work to render the frame and resumeFrame when the panel's work is complete. Without these calls, the message pipeline will assume the panel has completed the frame and will feed another frame to the panel to render. This additional frame will override the previous and result in dropped messages for the panel.

This change adds pauseFrame/resumeFrame logic to the PanelExtensionAdapter. As part of this change I've re-organized the rendering flow logic. The `queueRender` function is replaced with a `useLayoutEffect` that triggers when various inputs for renderState change. A new function `initRenderStateBuilder` returns a builder function that creates RenderState given these various inputs. This new builder functions is invoked inside of the layout effect and the panel's render function is called.

Fixes: #4091